+++ src/openeo_aggregator/constants.py
@@ -1,0 +1,6 @@
+JOB_OPTION_SPLIT_STRATEGY = "split_strategy"
+
+JOB_OPTION_TILE_GRID = "tile_grid"
+
+# Experimental feature to force a certain upstream back-end through job options
+JOB_OPTION_FORCE_BACKEND = "_agg_force_backend"